Add Haroun and the Sea of Stories by Salman Rushdie
Most might not consider this, science fiction, but all of the "magic" in the book is controlled by science.

P2C2Es(Processes too Complicated to Explain) and M2C2D(Machines too Complicated to Describe) are behind all the "magic" and are under the supervision of the Eggheads of Earth's second moon, Kahani.