Added my commentary on a couple of movies
I added the commentary to the movies I watched in the last couple of months: Transcendence, Her, Lucy, Ex Machina, Chappie and Interstellar.

I will not submit the patch directly to the master branch so other contributors can say their opinion about my comments. They might be a bit subjective for this list.