Add `Edge of Tomorrow` (2014) to the Movies section (fixed #105)
Not gonna merge the pull request directly so someone can proof read it and approve it if I did everything right.

I didn't follow the comments directly from the issue. I put my comment on the second place because I was referencing to the first comment when I mentioned Groundhog Day. I hope that's okay with you.