Removed Soft Science Fiction and Merged Fantasy categories
Soft Science Fiction is ambiguous; most of the novels here could fit in it. When you take into account that it is just the opposite of Hard Science Fiction, which really ought to be a category in my opinion, it was a useless category that was just sucking stuff away from the default category without real reason. I think it is better removed.

Low and High fantasy, as well, is not a useful distinction when there is only one author, and only three examples in both categories. I just merged them into one.