Add House of Leaves (#269)
I can't say I've ever been "excited" to read a book, but after hearing about _House of Leaves_ I was undeniably excited. I was in the middle of another book when my wife surprised me with this one, and I finished the book I was already reading much more quickly than I normally would have just so I could start reading _House of Leaves_.

If you're looking for a book which lays out all the answers for you and packages things up neatly with a bow, this is not for you 😄

---

This book is widely considered "horror" as well but I wasn't sure if a new "Horror" subsection should be created, so I put the entry in the general "Science Fiction" section. If you'd like a "Horror" subsection created as part of this PR, let me know and I can add it and move _House of Leaves_ there!