Nested Remembrance of Earth's Past Trilogy books
It looks like the last PR organized all titles by A-Z, but then any books in a series are no longer proximate to each other as one reads. I think a better organization scheme would be to have standalone books and series titles be A-Z as H4s but then nest the books in a series under the series title as H5s.